+++ test/jest/controls/control.selectMany.test.js
@@ -89,9 +89,16 @@ describe("control.selectMany", () => {
       `"<ul id="txt3" role="listbox" aria-label="Items" tabindex="-1" aria-multiselectable="true"><li role="option" style="display: none;">Donny</li><li role="option" aria-selected="false" id="txt4" style="display: none;">Mikky</li><li role="option" style="display: none;">Leo</li><li role="option" style="">Splinter</li></ul>"`
     );
 
+    el.$value = undefined;
+    await h.wait();
+    await h.userClick(el.$refPopup.querySelectorAll("li")[1]); // select 1st visible
+    expect(el.$value).toStrictEqual([20]);
+
+    el.$value = [10, 30, 20];
+    await h.wait(1);
     await h.userTypeText(el.$refInput, "s");
     expect(el.$refPopup.innerHTML).toMatchInlineSnapshot(
-      `"<ul id="txt3" role="listbox" aria-label="Items" tabindex="-1" aria-multiselectable="true"><li role="option" style="display: none;">Donny</li><li role="option" aria-selected="false" id="txt4" style="display: none;">Mikky</li><li role="option" style="display: none;">Leo</li><li role="option" style="" aria-selected="false" id="txt6" focused="">Splinter</li></ul>"`
+      `"<ul id="txt3" role="listbox" aria-label="Items" tabindex="-1" aria-multiselectable="true"><li role="option" style="display: none;">Donny</li><li role="option" aria-selected="false" id="txt4" style="display: none;">Mikky</li><li role="option" style="display: none;">Leo</li><li role="option" style="" aria-selected="false" id="txt8" focused="">Splinter</li></ul>"`
     );
     el.$refInput.dispatchEvent(new KeyboardEvent("keydown", { key: "Enter", bubbles: true, cancelable: true }));
     await h.wait();
@@ -99,7 +106,7 @@ describe("control.selectMany", () => {
     expect(el.$refInput.value).toBe("");
     expect(el.$isOpen).toBe(false); // because no-items anymore
     expect(el.$refPopup.innerHTML).toMatchInlineSnapshot(
-      `"<ul id="txt3" role="listbox" aria-label="Items" tabindex="-1" aria-multiselectable="true"><li role="option" style="display: none;">Donny</li><li role="option" aria-selected="false" id="txt4" style="display: none;">Mikky</li><li role="option" style="display: none;">Leo</li><li role="option" style="display: none;" aria-selected="false" id="txt6">Splinter</li><li role="option" aria-disabled="true" aria-selected="false">No Items</li></ul>"`
+      `"<ul id="txt3" role="listbox" aria-label="Items" tabindex="-1" aria-multiselectable="true"><li role="option" style="display: none;">Donny</li><li role="option" aria-selected="false" id="txt4" style="display: none;">Mikky</li><li role="option" style="display: none;">Leo</li><li role="option" style="display: none;" aria-selected="false" id="txt8">Splinter</li><li role="option" aria-disabled="true" aria-selected="false">No Items</li></ul>"`
     );
 
     // expected no-items
@@ -107,10 +114,10 @@ describe("control.selectMany", () => {
     await h.wait();
     expect(el.$isOpen).toBe(true);
     expect(el.$refInput.outerHTML).toMatchInlineSnapshot(
-      `"<input placeholder=" " type="text" id="txt1" role="combobox" aria-haspopup="listbox" aria-expanded="true" autocomplete="off" aria-autocomplete="list" class="" aria-owns="txt3" aria-controls="txt3" aria-activedescendant="txt8">"`
+      `"<input placeholder=" " type="text" id="txt1" role="combobox" aria-haspopup="listbox" aria-expanded="true" autocomplete="off" aria-autocomplete="list" class="" aria-owns="txt3" aria-controls="txt3" aria-activedescendant="txt10">"`
     );
     expect(el.$refPopup.innerHTML).toMatchInlineSnapshot(
-      `"<ul id="txt3" role="listbox" aria-label="Items" tabindex="-1" aria-multiselectable="true"><li role="option" style="display: none;" aria-selected="false" id="txt8" focused="">Donny</li><li role="option" aria-selected="false" id="txt4" style="display: none;">Mikky</li><li role="option" style="display: none;">Leo</li><li role="option" style="display: none;" aria-selected="false" id="txt6">Splinter</li><li role="option" aria-disabled="true" aria-selected="false">No Items</li></ul>"`
+      `"<ul id="txt3" role="listbox" aria-label="Items" tabindex="-1" aria-multiselectable="true"><li role="option" style="display: none;" aria-selected="false" id="txt10" focused="">Donny</li><li role="option" aria-selected="false" id="txt4" style="display: none;">Mikky</li><li role="option" style="display: none;">Leo</li><li role="option" style="display: none;" aria-selected="false" id="txt8">Splinter</li><li role="option" aria-disabled="true" aria-selected="false">No Items</li></ul>"`
     );
 
     await h.userClick(el.$refItems[0]); // remove item by click
@@ -118,7 +125,7 @@ describe("control.selectMany", () => {
     expect(el.$isOpen).toBe(true);
     expect(el.$value).toStrictEqual([30, 20, 40]); // because selected Splinter
     expect(el.$refPopup.innerHTML).toMatchInlineSnapshot(
-      `"<ul id="txt3" role="listbox" aria-label="Items" tabindex="-1" aria-multiselectable="true"><li role="option" style="" aria-selected="false" id="txt8" focused="">Donny</li><li role="option" aria-selected="false" id="txt4" style="display: none;">Mikky</li><li role="option" style="display: none;">Leo</li><li role="option" style="display: none;" aria-selected="false" id="txt6">Splinter</li><li role="option" aria-disabled="true" aria-selected="false" style="display: none;">No Items</li></ul>"`
+      `"<ul id="txt3" role="listbox" aria-label="Items" tabindex="-1" aria-multiselectable="true"><li role="option" style="" aria-selected="false" id="txt10" focused="">Donny</li><li role="option" aria-selected="false" id="txt4" style="display: none;">Mikky</li><li role="option" style="display: none;">Leo</li><li role="option" style="display: none;" aria-selected="false" id="txt8">Splinter</li><li role="option" aria-disabled="true" aria-selected="false" style="display: none;">No Items</li></ul>"`
     );
 
     el.$options.disabled = true;
@@ -141,7 +148,7 @@ describe("control.selectMany", () => {
       .spyOn(window, "matchMedia")
       .mockImplementationOnce((s) => ({ matches: s !== "(hover: hover) and (pointer: fine)" })); // simulate touchscreen
     expect(el.$refInput.parentElement.innerHTML).toMatchInlineSnapshot(
-      `"<span item="" aria-hidden="true">Leo</span><span item="" aria-hidden="true">Mikky</span><span item="" aria-hidden="true">Splinter</span><input placeholder=" " type="text" id="txt1" role="combobox" aria-haspopup="listbox" aria-expanded="false" autocomplete="off" aria-autocomplete="list" class="wup-hidden" aria-owns="txt10" aria-controls="txt10"><strong></strong>"`
+      `"<span item="" aria-hidden="true">Leo</span><span item="" aria-hidden="true">Mikky</span><span item="" aria-hidden="true">Splinter</span><input placeholder=" " type="text" id="txt1" role="combobox" aria-haspopup="listbox" aria-expanded="false" autocomplete="off" aria-autocomplete="list" class="wup-hidden" aria-owns="txt12" aria-controls="txt12"><strong></strong>"`
     );
     const item = el.$refItems[0];
     await h.userTap(item);
@@ -150,14 +157,14 @@ describe("control.selectMany", () => {
     expect(el.$isOpen).toBe(true); // 1st tap only to focus control
     // item must be not-removed
     expect(el.$refInput.parentElement.innerHTML).toMatchInlineSnapshot(
-      `"<span item="" aria-hidden="true">Leo</span><span item="" aria-hidden="true">Mikky</span><span item="" aria-hidden="true">Splinter</span><input placeholder=" " type="text" id="txt1" role="combobox" aria-haspopup="listbox" aria-expanded="true" autocomplete="off" aria-autocomplete="list" class="" aria-owns="txt13" aria-controls="txt13"><strong></strong>"`
+      `"<span item="" aria-hidden="true">Leo</span><span item="" aria-hidden="true">Mikky</span><span item="" aria-hidden="true">Splinter</span><input placeholder=" " type="text" id="txt1" role="combobox" aria-haspopup="listbox" aria-expanded="true" autocomplete="off" aria-autocomplete="list" class="" aria-owns="txt15" aria-controls="txt15"><strong></strong>"`
     );
 
     await h.userTap(item);
     await h.wait();
     // 2nd tap removes item
     expect(el.$refInput.parentElement.innerHTML).toMatchInlineSnapshot(
-      `"<span item="" aria-hidden="true">Mikky</span><span item="" aria-hidden="true">Splinter</span><input placeholder=" " type="text" id="txt1" role="combobox" aria-haspopup="listbox" aria-expanded="true" autocomplete="off" aria-autocomplete="list" class="" aria-owns="txt13" aria-controls="txt13"><strong></strong>"`
+      `"<span item="" aria-hidden="true">Mikky</span><span item="" aria-hidden="true">Splinter</span><input placeholder=" " type="text" id="txt1" role="combobox" aria-haspopup="listbox" aria-expanded="true" autocomplete="off" aria-autocomplete="list" class="" aria-owns="txt15" aria-controls="txt15"><strong></strong>"`
     );
 
     // input hidden if readOnlyInput and values exist
@@ -166,9 +173,23 @@ describe("control.selectMany", () => {
     el.focus();
     await h.wait(1);
     expect(el.$refInput.outerHTML).toMatchInlineSnapshot(
-      `"<input placeholder=" " type="text" id="txt1" role="combobox" aria-haspopup="listbox" aria-expanded="true" autocomplete="off" class="wup-hidden" aria-owns="txt16" aria-controls="txt16" aria-describedby="txt15" readonly="">"`
+      `"<input placeholder=" " type="text" id="txt1" role="combobox" aria-haspopup="listbox" aria-expanded="true" autocomplete="off" class="wup-hidden" aria-owns="txt18" aria-controls="txt18" aria-describedby="txt17" readonly="">"`
     );
     el.$options.readOnlyInput = false;
+
+    await h.userTypeText(el.$refInput, "Mik");
+    // simulate blur-by outside click
+    document.body.dispatchEvent(new MouseEvent("mousedown", { cancelable: true, bubbles: true }));
+    el.blur();
+    await h.wait(100);
+    expect(el.$refInput.value).toBe("Mik");
+    document.body.dispatchEvent(new MouseEvent("mouseup", { cancelable: true, bubbles: true }));
+    document.body.dispatchEvent(new MouseEvent("click", { cancelable: true, bubbles: true }));
+    document.body.focus();
+    await h.wait();
+    expect(el.$refInput.value).toBe(" ");
+
+    expect(el.clearFilterMenuItems).not.toThrow(); // for coverage
   });
 
   test("animation for removed item", async () => {